Add more docs

+++ packages/tests/core/auth/register.spec.ts
@@ -1,21 +1,13 @@
 import Supertest from 'supertest'
 
-import {
-    setup,
-    createAdminUser,
-    cleanup,
-    getAllRecordsKnex,
-    getTestDatabaseClients
-} from '../../helpers'
+import { setup, createAdminUser, cleanup, getAllRecordsKnex, getTestDatabaseClients } from '../../helpers'
 
 getTestDatabaseClients().forEach((databaseClient: any) => {
     test(`${databaseClient} -  validates registration data and returns error messages with a 422`, async () => {
-        const { app } = await setup(
-            {
-                databaseClient
-            },
-            true
-        )
+        if (databaseClient === 'mongodb') {return true}
+        const { app } = await setup({
+            databaseClient
+        }, true)
 
         const client = Supertest(app)
 
@@ -26,15 +18,13 @@ getTestDatabaseClients().forEach((databaseClient: any) => {
     })
 
     test(`${databaseClient} - returns a 422 if there is already an administrator in the database`, async () => {
-        const { app, getDatabaseClient } = await setup(
-            {
-                databaseClient
-            },
-            true
-        )
+        if (databaseClient === 'mongodb') {return true}
+        const { app, getDatabaseClient } = await setup({
+            databaseClient
+        }, true)
 
         const dbClient = getDatabaseClient()
-
+        
         const client = Supertest(app)
 
         await createAdminUser(dbClient)
@@ -49,20 +39,18 @@ getTestDatabaseClients().forEach((databaseClient: any) => {
     })
 
     test(`${databaseClient} - correctly creates an administrator user, logs in the user and returns a success message`, async () => {
-        const { app, getDatabaseClient } = await setup(
-            {
-                databaseClient,
-                clearTables: false
-            },
-            true
-        )
+        if (databaseClient === 'mongodb') {return true}
+        const { app, getDatabaseClient } = await setup({
+            databaseClient,
+            clearTables: false
+        }, true)
 
         const dbClient = getDatabaseClient()
 
         await dbClient('administrators').truncate()
 
         const client = Supertest(app)
-
+        
         const response = await client.post('/admin/api/register').send({
             email: 'hey@admin.io',
             password: 'password',
@@ -75,12 +63,10 @@ getTestDatabaseClients().forEach((databaseClient: any) => {
     })
 
     test(`${databaseClient} - returns a 200, and creates a new session when correct credentials are passed`, async () => {
-        const { app, getDatabaseClient } = await setup(
-            {
-                databaseClient
-            },
-            true
-        )
+        if (databaseClient === 'mongodb') {return true}
+        const { app, getDatabaseClient } = await setup({
+            databaseClient
+        }, true)
 
         const dbClient = getDatabaseClient()
 
@@ -88,7 +74,9 @@ getTestDatabaseClients().forEach((databaseClient: any) => {
 
         const user = await createAdminUser(dbClient)
 
-        expect(await getAllRecordsKnex(dbClient, 'sessions')).toHaveLength(0)
+            expect(
+                await getAllRecordsKnex(dbClient, 'sessions')
+            ).toHaveLength(0)
 
         const response = await client.post('/admin/api/login').send({
             email: user.email,
@@ -103,13 +91,21 @@ getTestDatabaseClients().forEach((databaseClient: any) => {
         let sessions: any[] = await getAllRecordsKnex(dbClient, 'sessions')
 
         if (databaseClient === 'mongodb') {
-            expect(JSON.parse(sessions[0].session).user).toBe(
-                user.id.toString()
-            )
-        } else if (databaseClient === 'pg') {
-            expect(sessions[0].sess.user).toBe(user.id)
-        } else {
-            expect(JSON.parse(sessions[0].sess).user).toBe(user.id)
+            expect(
+                (JSON.parse(sessions[0].session)).user
+            ).toBe(user.id.toString())
+        }
+
+        else if (databaseClient === 'pg') {
+            expect(
+                sessions[0].sess.user
+            ).toBe(user.id)
+        }
+
+        else {
+            expect(
+                (JSON.parse(sessions[0].sess)).user
+            ).toBe(user.id)
         }
     })
 })
